// src/types/address.ts
export interface CtreateAddress {
  firstname: string;
  lastname: string;
  phone: number;
  plaque: number;
  city: string;
  address: string;
}

export interface AddressType {
  _id: string;
  firstname: string;
  lastname: string;
  phone: number;
  plaque: number;
  city: string;
  address: string;
  createdAt: string;
  updatedAt: string;
}